==> 3. Clustering.ipynb <==
from Scripts.user_analysis import user
from Scripts.choose_random_user import randUsr
from matplotlib import pyplot as plt
import os
import pandas as pd
import random


#thedir = '/Users/okkevanderwal/Downloads/Geolife Trajectories 1.3/Data'
thedir = '.../Geolife Trajectories 1.3/Data'
dirs = [name for name in os.listdir(thedir) if os.path.isdir(os.path.join(thedir, name))]

def clustering(root):    
    dirs = [name for name in os.listdir(root) if os.path.isdir(os.path.join(root, name))]
    for x in dirs: 
        id_user = dirs[int(x)]
        path = str(root + '/' + dirs[int(x)])
        
        user_x = user(path)
        array = user_x.clusterLocs()
        cluster_mean = array.mean()        
        dict_cluster = {id_user:cluster_mean}
        
        print(dict_cluster)
        
clustering(thedir)

df_cluster = pd.DataFrame(dict_cluster)

df.to_csv('Final_CSV/Clusters.csv')
