==> 4. Merging Clustering and Features.ipynb <==
import pandas as pd
import numpy as np

seed = 20

df_features = pd.read_csv('Final_CSV/trajFeatures.csv')

# Read CSV file with Clustering file
df_clusters = pd.read_csv('Final_CSV/Clusters.csv')
df_clusters.head(5)

# Merge Dataframe with Clustering file
df_features['user_id'] = df_features['Path'].astype(str).str[0:3]
df_features['user_id'] = df_features['user_id'].astype(int)
df_features.head(10)

df = pd.merge(df_features, df_clusters, on='user_id')

df.shape

df.head(10)

df.to_csv('Final_CSV/Geolife_Features.csv')
