==> 7. Random_User_Analysis.ipynb <==
from Scripts.user_analysis import user
from Scripts.choose_random_user import randUsr
from matplotlib import pyplot as plt
import os

thedir = '.../Geolife Trajectories 1.3/Data'
dirs = [name for name in os.listdir(thedir) if os.path.isdir(os.path.join(thedir, name))]

### Random User

#u2 = randUsr(thedir) #for random user
u2 = '/Users/okkevanderwal/Downloads/Geolife Trajectories 1.3/Data/170' #for specific user
u3 = user(u2)
print(u2,u3)

u3.plotClusters()

array = u3.clusterLocs()
array.mean()

u3.plotAllTrajs()
plt.rcParams["figure.figsize"] = [8,8]
plt.xlim(-5000,5000)
plt.ylim(-5000,5000)
plt.show()

u3.startEnd_heatmap()
plt.rcParams["figure.figsize"] = [8,8]
plt.xlim(-5000,5000)
plt.ylim(-5000,5000)
plt.show()

u3.modeInfo()

Their activity over the full range of recording:

u3.dates_hist()
plt.rcParams["figure.figsize"] = [8,4]
plt.show()
